feat: Add initial core models, API routes, and repository pattern for user and head of family entities.

+++ database/migrations/2026_03_16_035529_create_head_of_families_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
 
             $table->string('profile_picture')->nullable();
-            $table->integer('identity_number');
+            $table->bigInteger('identity_number');
             $table->enum('gender', ['male', 'female']);
             $table->date('date_of_birth');
             $table->string('phone_number');
